+++ pelican-core/src/test/kotlin/io/github/matthewjones372/pelican/NegotiationTest.kt
@@ -1,5 +1,6 @@
 package io.github.matthewjones372.pelican
 
+import io.github.matthewjones372.pelican.spi.*
 import io.kotest.assertions.withClue
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test